+++ website/package.json
@@ -19,7 +19,6 @@
     "@babel/core": "^7.26.0",
     "@babel/plugin-syntax-jsx": "^7.25.9",
     "@mdx-js/mdx": "^3.1.0",
-    "@swc/core": "^1.10.1",
     "@types/babel__standalone": "^7.1.9",
     "@types/node": "^18.11.17",
     "prettier": "^3.4.2",
